Move state management to its own module
